+++ lib/src/dart/extensions/src/datetime_extension.dart
@@ -20,91 +20,92 @@ extension DateTimeExtension on DateTime {
   ///
   /// Example usage:
   /// DateTime now = DateTime.now();
-  /// String formattedDate = now.format('dd/MM/yyyy', monthFormat: 'MMM');
-  String format(String format, {String? monthFormat, String? dayFormat}) {
-    Map<int, String> monthsShort = {
-      1: 'Jan',
-      2: 'Feb',
-      3: 'Mar',
-      4: 'Apr',
-      5: 'May',
-      6: 'Jun',
-      7: 'Jul',
-      8: 'Aug',
-      9: 'Sep',
-      10: 'Oct',
-      11: 'Nov',
-      12: 'Dec'
+  /// String formattedDate = now.format('dd/MM/yyyy');
+  String format(String format) {
+    final Map<String, String> shortMonths = {
+      '01': 'Jan',
+      '02': 'Feb',
+      '03': 'Mar',
+      '04': 'Apr',
+      '05': 'May',
+      '06': 'Jun',
+      '07': 'Jul',
+      '08': 'Aug',
+      '09': 'Sep',
+      '10': 'Oct',
+      '11': 'Nov',
+      '12': 'Dec'
     };
-    Map<int, String> monthsFull = {
-      1: 'January',
-      2: 'February',
-      3: 'March',
-      4: 'April',
-      5: 'May',
-      6: 'June',
-      7: 'July',
-      8: 'August',
-      9: 'September',
-      10: 'October',
-      11: 'November',
-      12: 'December'
+
+    final Map<String, String> fullMonths = {
+      '01': 'January',
+      '02': 'February',
+      '03': 'March',
+      '04': 'April',
+      '05': 'May',
+      '06': 'June',
+      '07': 'July',
+      '08': 'August',
+      '09': 'September',
+      '10': 'October',
+      '11': 'November',
+      '12': 'December'
     };
-    Map<int, String> daysShort = {
-      1: 'Mon',
-      2: 'Tue',
-      3: 'Wed',
-      4: 'Thu',
-      5: 'Fri',
-      6: 'Sat',
-      7: 'Sun'
+
+    final Map<String, String> shortDays = {
+      '1': 'Mon',
+      '2': 'Tue',
+      '3': 'Wed',
+      '4': 'Thu',
+      '5': 'Fri',
+      '6': 'Sat',
+      '7': 'Sun'
     };
-    Map<int, String> daysFull = {
-      1: 'Monday',
-      2: 'Tuesday',
-      3: 'Wednesday',
-      4: 'Thursday',
-      5: 'Friday',
-      6: 'Saturday',
-      7: 'Sunday'
+
+    final Map<String, String> fullDays = {
+      '1': 'Monday',
+      '2': 'Tuesday',
+      '3': 'Wednesday',
+      '4': 'Thursday',
+      '5': 'Friday',
+      '6': 'Saturday',
+      '7': 'Sunday'
     };
-    String result = format.replaceAllMapped(
-        RegExp(r'(dd|MM|MMM|MMMM|yyyy|HH|hh|mm|ss|a|EEE|EEEE)'), (Match m) {
-      switch (m.group(0)) {
-        case 'dd':
-          return this.day.toString().padLeft(2, '0');
-        case 'MM':
-          return monthFormat == 'MMM'
-              ? monthsShort[this.month]! // Short month name: Jan-Dec
-              : this.month.toString().padLeft(2, '0'); // Numeric month: 01-12
-        case 'MMM':
-          return monthsShort[this.month]!; // Short month name: Jan-Dec
-        case 'MMMM':
-          return monthsFull[this.month]!; // Full month name: January-December
-        case 'yyyy':
-          return this.year.toString();
-        case 'HH':
-          return this.hour.toString().padLeft(2, '0');
-        case 'hh':
-          int hour = this.hour > 12 ? this.hour - 12 : this.hour;
-          return hour.toString().padLeft(2, '0');
-        case 'mm':
-          return this.minute.toString().padLeft(2, '0');
-        case 'ss':
-          return this.second.toString().padLeft(2, '0');
-        case 'a':
-          return this.hour < 12 ? 'AM' : 'PM';
-        case 'EEE':
-          return dayFormat == 'EEE'
-              ? daysShort[this.weekday]! // Short day name: Mon-Sun
-              : daysFull[this.weekday]!; // Full day name: Monday-Sunday
-        case 'EEEE':
-          return daysFull[this.weekday]!; // Full day name: Monday-Sunday
-        default:
-          return m.group(0)!;
-      }
-    });
-    return result;
+
+    return format.replaceAllMapped(
+      RegExp(r'(dd|MM|MMM|MMMM|yyyy|HH|hh|mm|ss|a|EEE|EEEE|[^\w])'),
+      (Match m) {
+        switch (m.group(0)) {
+          case 'dd':
+            return this.day.toString().padLeft(2, '0');
+          case 'MM':
+            return this.month.toString().padLeft(2, '0');
+          case 'MMM':
+            return shortMonths[this.month.toString().padLeft(2, '0')]!;
+          case 'MMMM':
+            return fullMonths[this.month.toString().padLeft(2, '0')]!;
+          case 'yyyy':
+            return this.year.toString();
+          case 'HH':
+            return this.hour.toString().padLeft(2, '0');
+          case 'hh':
+            int hour = this.hour > 12 ? this.hour - 12 : this.hour;
+            return hour.toString().padLeft(2, '0');
+          case 'mm':
+            return this.minute.toString().padLeft(2, '0');
+          case 'ss':
+            return this.second.toString().padLeft(2, '0');
+          case 'a':
+            return this.hour < 12 ? 'AM' : 'PM';
+          case 'EEE':
+            return shortDays[this.weekday.toString()]!;
+          case 'EEEE':
+            return fullDays[this.weekday.toString()]!;
+          default:
+            return m.group(0)!;
+        }
+      },
+    );
   }
 
   /// Returns true if the DateTime object represents a date in the past.