+++ lib/src/dart/extensions/src/string_extension.dart
@@ -12,10 +12,20 @@ extension StringExtension on String {
   ///Checks [string] is Alpha Numerical or not and return [bool]
   bool get isAlphaNumeric => regex(r'^[a-zA-z0-9]+$');
 
-  ///Checks [string] is valid email or not and return [bool]
-  bool get isEmail => regex(
-        r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+",
+  /// Checks if string is URL.
+  bool get isURL => regex(
+        r"^((((H|h)(T|t)|(F|f))(T|t)(P|p)((S|s)?))\://)?(www.|[a-zA-Z0-9].)[a-zA-Z0-9\-\.]+\.[a-zA-Z]{2,7}(\:[0-9]{1,5})*(/($|[a-zA-Z0-9\.\,\;\?\'\\\+&amp;%\$#\=~_\-]+))*$",
       );
+
+  /// Checks if string is email.
+  bool get isEmail => regex(
+        r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$',
+      );
+
+  /// Checks if string is boolean.
+  static bool isBool(String value) {
+    return value == "true" || value == "false";
+  }
 
   ///Check this string min length, if satisfied condition then return true
   bool minLen(int min) => length >= min;